add system stock and buy them

+++ Models/AccountInfo.cs
@@ -3,7 +3,7 @@
     public class AccountInfo
     {
         public string name {  get; set; }
-        public int balance {  get; set; }
+        public float balance {  get; set; }
         public string uuid { get; set; }
     }
 }
